requirements.txt
openai
sqlalchemy==1.4.22
python-telegram-bot==13.7
# python-telegram-bot==13.4.1
psycopg2-binary
loguru
dateparser